+++ app_server/public/src/css/partials/_message.scss
@@ -1,4 +1,4 @@
-#error {
+#alert {
   background-color: $primary-color;
   margin: 0;
